Implement proper modifier resolution ordering
Summary: If a modifier for constraint A selects on modifier for constraint B, then all modifiers for constraint B must be resolved before modifiers of constraint A. This diff implements this proper ordering by traversing the modifiers to obtain a topologically sorted ordering of constraint settings. Then, we loop over this ordering to resolve modifiers.

Reviewed By: stepancheg

Differential Revision: D50206983

fbshipit-source-id: 5bda41e50d68a4cd70b422e2c7ba169726182fed