Change cfg_modifier to cfg_modifiers
Summary:
Now that we expect modifiers to be set as a single dictionary under key "buck.cfg_modifiers", change `cfg_modifier` function on target to `cfg_modifiers`. `cfg_modifiers` accepts a dictionary of constraint setting to modifier, as opposed to `cfg_modifier` which just accepts a single constraint setting and a single modifier.
Update all call sites to use the new function.

Reviewed By: stepancheg

Differential Revision: D51104066

fbshipit-source-id: c75fa7f7310abc876a54d8faa3ad31c2b19f91d6