Add conversion to and back from json
Summary: Add functions to convert modifiers (specifically everything that can go into `TaggedModifier` type) to and back from JSON. We need to do this because PACKAGE and target modifiers are stored in core as JSONs, but relying on core to convert for us is bad because converting to JSON is lossy. We need to manually convert to JSON before writing the modifiers and then convert them back to record types in cfg constructor.

Reviewed By: stepancheg

Differential Revision: D50702215

fbshipit-source-id: fd3067efa2a4872237058aa7f586a2d1ea7ade83