client: Errors with `--console none`
Summary:
Adds a new `ErrorConsole` which only prints errors at the end of the build, to be used by `--console none`.

The general idea was discussed in chat. This does not yet solve all of the problems with how things currently work, but it should at least be an improvement

Reviewed By: wendy728

Differential Revision: D51185986

fbshipit-source-id: 9818fa59a0f4938395ab1a20cb3b6c2503087191