Remove `cfg_modifiers` function
Summary: This function is not easily useable in practice because it expects all modifiers to be specified in one call to `cfg_modifiers`. In practice, there are going to be target modifiers added in different layers of macros, so it's not possible to expect all target modifiers to be passed in one place. Make target modifiers just directly accept a list of modifiers rather than `cfg_modifiers(<a list of modifiers>)`.

Reviewed By: wendy728

Differential Revision: D51264306

fbshipit-source-id: cfd58f2a9a06119c56c5d05e2368566f04909a2a