Accept list of modifiers instead of dictionary of constraint setting to modifier in cfg constructor
Summary:
Previously, PACKAGE modifiers and target modifiers need to be a dictionary of constraint setting to modifier, where the constraint setting matches the modifier. When used in practice, this requirement ends up being too cumbersome because a lot of times it's hard to remember the name of the constraint setting for a constraint value target.
This diff changes cfg constructor so that it accepts lists of modifiers instead of dictionary, while keeping the existing enforcement checks we have that a `modifier_select` can only modify a single constraint setting. `set_cfg_modifiers` and `cfg_modifiers` functions will be updated in future diffs.

Reviewed By: stepancheg

Differential Revision: D51213256

fbshipit-source-id: 70d8e699b8782c0b294e6e533e91d89807b44837