Accept variable-length list of modifiers instead of dictionary of constraint setting to modifier in set_cfg_modifiers and cfg_modifiers
Summary: Same as  D51213256 but for `set_cfg_modifiers` and `cfg_modifiers` functions. Requiring modifiers to specify constraint_setting is too cumbersome. This diff gets rid of this restriction of `set_cfg_modifiers` and `cfg_modifiers`. Now instead of accepting a dictionary as input, both functions accept an `*args` where each `arg` is a modifier.

Reviewed By: stepancheg

Differential Revision: D51219419

fbshipit-source-id: d24a55db1a60de757b048456cafffb880ea87ea6