Remove TaggedModifierInfo
Summary: `TaggedModifierInfo` was an object used to represent a resolved modifier tagged with the location it was set at (PACKAGE, target, or command line). However, we don't actually need it because the location tagging is only used for error messages, and by the point that this `TaggedModifierInfo` is created, we don't have user error messages anymore (only internal ones). This diff removes `TaggedModifierInfo`.

Reviewed By: wendy728

Differential Revision: D51208058

fbshipit-source-id: 4251dfe30e05d0c59a54e3cb879894b2a52fc107