Add materialize_failed_inputs flag to CommonBuildOptions
Summary:
This stack is to support materializing inputs for actions that failed on RE. It's controlled by a flag as part of the common build options, and currently it's only implemented for actions that ran on RE without remote dep cache or action cache, but I plan on adding support for that later once I make sure general implementation is correct.

Local actions already automatically materialize their inputs, so no need to touch that.

Reviewed By: scottcao

Differential Revision: D51111450

fbshipit-source-id: 0e708811d8e291ce081ee458b86ce67f673c3993