filewatcher: Report global rev of mergebase
Summary:
We currently report the mergebase as a commit hash. Commit hashes have limited usefullness in most contexts. Global revs are much better, primarily because they're totally ordered - they make it easy to ask "did this build include this diff or not."

Unfortunately, the only way to implement this is - as far as I can tell - to shell out to hg. Watchman doesn't provide this info. We can be pretty conservative about this though. We do at least know when the mergebase changed though, so we don't need to do this on every commad.

Gated behind a buckconfig, since non-hg repos can't make use of this anyway.

Reviewed By: iguridi

Differential Revision: D51244021

fbshipit-source-id: afc5095cbf7bd939774c64ce25e6012ed633bceb