Move kapt_base64_encoder to prelude/kotlin/tools
Summary: This diff stack moves the necessary kotlin toolchains from xplat to prelude/kotlin/tools because I am trying to onboard wa-android to kotlincd. Currently wa-android lives in its own repo and we sync buck2/prelude to that repo periodically. This essentially allows wa-android to have access to these jars or tools that kotlincd depends on.

Reviewed By: IanChilds

Differential Revision: D51293270

fbshipit-source-id: 46917285b57734c260f7600554393c23aa7b8d0c