Print the header when exporting to CSV for what-materialized and what-uploaded
Summary:
`what-materialized` and `what-uploaded` export to CSV but were missing to print the header, making it difficult to inspect the results. This diff fixes that.

Next diff D51117208 does the same for `what-ran`.

NB: Before we were creating `csv::Writer` for each row (without the headers), now we just create it once  with the default settings (which automatically prints the header and then serialize further. This version is more efficient I think.

Reviewed By: stepancheg

Differential Revision: D50646315

fbshipit-source-id: ecd71e5df241da8ada268b08f0c839af01180485