rename a function
Summary:
ezgicicek was asking some questions about this code and I found this a little
confusing. The name of the function "are we emitting this immediately?"
suggests that if the answer is "no" we might emit it later, but in reality it's
now or never.

Let's clarify that the point is that the action is done and the question is
do we want to emit this action that is done.

Reviewed By: samkevich

Differential Revision: D51424322

fbshipit-source-id: 759a915b015bb8116461f1146d13eb85d9355cf1