buck2: whatran make --failed and --incomplete exclusive
Summary:
Because you can't have both those things be true at the same time. Also, move
incomplete to a better place, it was in with the formatting options (which are
also used by the console), but it is only used by whatran so put it there
instead.

Note: I'd normally make this a group, but ezgicicek is about to add some more options
to this (like showing stderr) and those will also have conflicts with `--incomplete`
(but not with `--failed`) so let's get the right pattern in place for this now

Reviewed By: samkevich

Differential Revision: D51424419

fbshipit-source-id: 3645fbcb1c3fbf76413d742971a61206f913e458