fix test to check process group termination
Summary:
Originally test was implemented only for unix and checked that child process is killed.
In this [PR](https://github.com/facebook/buck2/pull/483) test was changed to use a file rather than the PID because Windows reuses PIDs a lot.  But it actually didn't work. The test always passes even if child process is still alive.
This diff returns check for PIDs to ensure that child process is killed. To mitigate PID reuse we now also check for parent PID.

Reviewed By: stepancheg

Differential Revision: D51399236

fbshipit-source-id: 98fe2e25eb639658c04ffd220fa78403cb6e8f4f