buck2: don't log "Command Failed" when the user stops reading stdout or interrupts
Summary:
It's kinda unpleasant to e.g. do `buck2 log whatran | head -n5` and to be told:

```
Command failed: Broken pipe (os error 32)
```

But also we used to do that a while ago and people complained about it, then we
stopped, and now we're doing it again. I think that behavior was changed by
accident in D50689224 (that diff's summary does not say anything about wanting
that behavior).

Reviewed By: samkevich

Differential Revision: D51560556

fbshipit-source-id: 5fab4e3b97947e85404b2b90f94ec56b73230b07