Always include typing manifest for `srcs` of python binary targets
Summary:
There's more context in [this post](https://fb.workplace.com/groups/python.builds/permalink/3590732654581484/).

Basically, Pyre-in-FBCode rely on source-db builds, which in turn relies on whether the python provider of a given python target includes a `src_types` manifest file or not. If the manifest exists in the target, we'll add the `srcs` of the corresponding rule into the sourcedb. Otherwise, we'll ignore the `srcs` fields.

The question here is whether the native `python_binary()` rule should include this `src_types` manifest or not. Previously, we do not include the manifest, which actually works well for FBCode because the fbcode `python_binary()` macro will actually generate a "companion" `python_library()` target which already includes the `src_types` manifest information for the binary. Unfortunately, the same "companion library generation" behavior is missing from non-fbcode python binary wrappers (e.g. `fb_python_binary()`), meaning that Pyre won't be able to see files that are specified in `srcs` of `fb_python_binary()` targets.

In other words, this is what we have today:
- FBCode `python_binary` = native `python_binary()` (without sourcedb) + native `python_library()` (with sourcedb)
- `fb_python_binary` = native `python_binary()` (without sourcedb)

What this diff attempted is to include `src_types` manifest into all native `python_binary()` rules. This makes it possible for `srcs` in `fb_python_binary()` targets to show up in sourcedb:
- FBCode `python_binary` = native `python_binary()` (with sourcedb) + native `python_library()` (with sourcedb)
- `fb_python_binary` = native `python_binary()` (with sourcedb)

The downside of this solution is that now in FBCode, we're going to get 2 (duplicated) manifests both on the binary and on the companion library. This is not a big problem for us for now, as our BXL script downstream is able to de-dup the corresponding maps. But it'll definitely put slightly more work on the downstream script. I've raised this issue on the original thread but until we get a consensus on what to do there, let's at least address the missing files issue in sourcedb first.

Reviewed By: zsol, inseokhwang

Differential Revision: D50347321

fbshipit-source-id: 21158a0e6a0ba421129b6cc5864c8ffbe43ec097