buck2_error: No support for inferring `Display` on `Path`
Summary: thiserror will let you format a `Path` with `{}` as if it implemented display; implementing that requires support from a helper trait defined outside of the macro crate; instead of dealing with that, let's just delete it. Writing `#[error("{}", path_field.display())]` is not that much work. Diff also removes the places in the codebase where we use this feature

Reviewed By: stepancheg

Differential Revision: D51453174

fbshipit-source-id: f58ed880f9c4f4fc7cc754c6a6f86b7c8be9e6b9