buck2_error: Make `buck2_error_derive_new` buildable
Summary:
So let me explain what this is about:

`buck2_error_derive` currently works by doing clever tricks to forward parts of its implementation to `thiserror`. This has worked so far, with some downsides. The primary one is that clever tricks are hard to maintain and review.

Still, I felt that the tradeoff favored that implementation strategy for a while; however, we've now started running into additional issues. Specifically, getting `thiserror` to have the right behavior around forwarding `provide` impls in `#[error(transparent)]` cases is hard. Also, using `self` in `#[error]` formatting doesn't work. Finally, there's interest in reducing the subset of the thiserror API that we actually use in buck2.

Together, I think that means that we're now better off just rebuilding buck2_error_derive from a fork of thiserror. More code, but it'll be a lot easier to make it do what we want.

As a part of the decision to fork, the calculus around what features to support also becomes a bit different. The previous diff removes support from this crate for `#[from]`, `#[backtrace]` and optional `#[source]` fields. The second and third are unused in buck2 - the first is used, but in verfy few places and can be easily reimplemented by hand, so it doesn't make sense to keep hundreds of lines of implementation code around in this crate to support it.

This diff just adds a targets file for now, to make `buck2_error_derive_new` sort of buildable.

I won't bother setting up a cargo build for this. Instead, I'm just planning to use this crate to do the re-implementation, and then "release" it in the form of an atomic commit that copies the implementation over

Reviewed By: stepancheg

Differential Revision: D51453176

fbshipit-source-id: 48827388e8c049dc97de8f064b9c4ce09e181a44