rules: sysroot via rustc shim in build scripts
Summary:
Instead of having a different compiler on the toolchain for buildscripts, generate a rustc shim which adds the appropriate dependency args to the rustc invocation.

This does mean that buildscripts which do not correctly pass `--target` and try to compile something will break, since the dependencies are not availabel for all targets. I think that's fine. `--version` and similar still works of course

Reviewed By: shayne-fletcher

Differential Revision: D51461803

fbshipit-source-id: 42d4446143801be1f55be79464aa6946a95fee66