oss: Pin ring
Summary: The fixups for this are currently 0.17.5 specific. I could do the ring upgrade to 0.17.6, but I don't really see the point in that. There's no desperate need to be on the latest version, so we shouldn't make someone go through the pain of regenerating the headers each and every time ring releases a new minor version. Pin it for now, and then we'll unping it once we actually need to

Reviewed By: dtolnay

Differential Revision: D51647046

fbshipit-source-id: 9bdea1b99630cb9aa70b54d0261864c836afbe21