errors: Prepare derive for allowing `starlark::Result`
Summary:
When Stiopa and I initially discussed this, we weren't sure what kind of return types we want to allow in module/global builder derives. Clearly we'll have to start allowing `starlark::Result`. I propose that we additionally keep allowing `anyhow::Result` too, but only those two, ie with no intent to extend this further in the future.

I think the additional benefit of trying to ban `anyhow::Result` would be either zero or negative. On the one hand, it would require migrating a lot of code. But also, the ergonomics of `starlark::Error` will always be worse than `anyhow::Error` (since it's not `From<E: Error>`), and 99% of the time users will just want to create an `ErrorKind::Other` anyway.

This diff modifies the derive macro to syntactically allow `starlark::Result`, but it does not actually work yet since `StarlarkValue::invoke` still expects a `anyhow::Result`. TODOs are left in for the two places that need to be updated to "turn it on" so to speak, which will be done in-sync with a larger diff

Reviewed By: stepancheg

Differential Revision: D51573527

fbshipit-source-id: 16373d749c5eac947fb2ef91f005c51a6b7778bb