Disable incremental
Summary:
There's two reasons that we should do this. The first is that this hurts incremental rebuilds (in the buck sense, not the Rust sense), explained in  https://fb.workplace.com/groups/buck2dev/permalink/3584203148534398/ .

The second reason is that this also hurts initial startup times. Actions that are run locally aren't uploaded to the action cache (obvious security reasons). This means that the first build after a rebase ends up being a local only, no cache build, which is attrociously slow. The only reason we get any cache hits at all is because third party libraries ignore the incremental setting:

https://www.internalfb.com/code/fbsource/[4f4730cc83446645c0fa94e59c562a13828810be]/tools/build_defs/fbcode_macros/build_defs/lib/rust_common.bzl?lines=547

The actual numbers here are that the v2 isolation dir sees 96% cache hits, while the r-a isolation dir gets 70%.

Both of these things could be individually fixed via different means. I'm happy to provide some advice on how to do that, but don't have the bandwidth myself.

Reviewed By: davidbarsky

Differential Revision: D51697187

fbshipit-source-id: 9b4c2f13e9f38bc284abd1709a722a935e3fc52e