Add executable's mandatory runtime dependencies to DistInfo provider
Summary:
The current use of `DefaultInfo.other_outputs` by `rust_binary`, `cxx_binary`, and `go_binary` is conflating 2 different kinds of things:

- **Artifacts which are always required for running the executable.** These would need to be materialized regardless of whether the buck2 user plans to run the executable from their own shell or whether remote execution needs to run the executable as a dependency of another rule. This includes shared libraries that the executable dynamically links against, and resources.

- **Artifacts which are required for debugging the executable** (in the case of Split DWARF). Remote execution should never materialize external debuginfo because it's not running executables through a debugger. However, when a buck2 user builds an executable as the output of a build, they want to get the external debuginfo so that the executable is immediately usable in a debugger without needing to build some separate target to materialize the external debuginfo locally.

This diff adds a field called `nondebug_runtime_files` to `DistInfo`, to hold just the 1st category. This category is already referred to as "runtime files" throughout the prelude.

`DefaultInfo.other_outputs` will continue to be used for the union of both categories. This is everything you want to get when you do `buck2 build :main`.

By itself, this diff does not have any effect. I have confirmed that `DefaultInfo.other_outputs` and `DistInfo.runtime_files` always have the same contents everywhere that `DistInfo` appears in the result of a rule. Later diffs in this stack begin to diverge the contents appropriately.

Reviewed By: JakobDegen

Differential Revision: D51222272

fbshipit-source-id: a0c38f1428a1347e8ea140abb9096f1d82116e8c