errors: Return `DefError` from a function
Summary: I want to have an `ErrorKind::Internal(String)` variant. That works fine in principle; however, when we currently convert a `DefError` into a `starlark::Error`, we do not yet know if it will be an internal error - in some codepaths that error will later be converted into an internal error, and in some it won't. So delay the conversion until we know what to do with it. Use that in two diffs

Reviewed By: stepancheg

Differential Revision: D51735673

fbshipit-source-id: 403ffd02a55605da5e8ec167569aee969b1127d6