prelude: Set `native_unbundle_deps` correctly in prebuilt Rust libraries
Summary: Native unbundle deps doesn't only affect how a target is linked when it's depended on directy by a native rule. It also affects whether the target is introduced into the native link graph at all - introducing it without enabling the rest of the native unbundle deps behavior is wrong.

Reviewed By: capickett

Differential Revision: D51782341

fbshipit-source-id: 93a37a9943cdc958f9c79c5976855f90b2c89618