RePlatformFields
Summary:
Wrapper for `SortedMap<String, String>`.

Use it in `CacheUploader`.

Need it later to cache permissions by platform, and `RE::Platform` is not `Hash`. It is possible to derive `Hash` in generated proto files, but this approach is cleaner.

Also make it `Arc` to make it cheap when it is used as map key.

Reviewed By: wendy728

Differential Revision: D51688876

fbshipit-source-id: 11bb40599bc48b1d52f084ded11d702896e40029