accept declared artifacts in get_path_without_materialization
Summary: I don't see why we can't support this. The error message from passing in a declared artifact into this also sucks since the underlying rust types are different but the starlark type repr is the same ("expected `artifact`, found `artifact`"), so let's just support declared artifact in this function.

Reviewed By: scottcao

Differential Revision: D51808151

fbshipit-source-id: 76ea9089736a556b78851257a444972815826e61