buck2: Report CommandExecutionKind in action errors
Summary:
This is an oooooold SEV followup that never actually got done (T151398956).

tl;dr During this SEV, we saw actions failing during `materialize_inputs` but it was unclear whether these were local or remote actions.

This plumbs through an executor kind through _most_ (but not all) places we surface errors during action execution - in both local, remote, caching, and worker.

The places where we don't are marked with TODOs, mostly because I'm not sure which executor we've selected at that point. In some cases, we don't pick an executor til later (hence the optionality of executor kind in the error case).

Reviewed By: krallin

Differential Revision: D51519498

fbshipit-source-id: e162344e3dce551a408aac0bd4a90b70308ca131