buck2_error: Simplify a base case
Summary:
This stack:

Consider an error like
```
let e: buck2_error::Error anyhow::Error::new(MyType).context(Ctx).into();
```

In the past, `buck2_error` needed to support potentially downcasting to either `Ctx` or `MyType`, which meant that the root needed to store the `anyhow::Error` as is. That makes things pretty awkward, because it means that the error "root" might actually have a bunch more `source()`s hidden behind it. It results in complications in the formatting code as well as the error recovery code.

With downcasting gone, we can simplify everything by removing the need for all that. Instead, when we create the `buck2_error` above, we'll just walk the source chain all the way to the end and turn everything into a string as we go. We'll also always create the error root from whatever is at the bottom of the source chain

This diff:

Gets part of the way there

Reviewed By: stepancheg

Differential Revision: D52022075

fbshipit-source-id: 9100e08b0f41905b10de3566a3073a08da1877ab