add RE dependencies as part of CommandExecutor
Summary:
## Context

We want to have [license aware scheduling](https://docs.google.com/document/d/1LpVDp90x-II_a7akOqdtoMeFLRXA-hOUxY5DCSckaWQ/edit#heading=h.f80kk7jqstkf) in RE for ASIC workloads.

The idea is that a TARGET can specify "I need a certain dependency X to be acquired before running on RE" and RE will know how to acquire X and will wait to schedule the action until X is acquired.

## Diff

Accept a new parameter for RE dependencies as part of the `CommandExecutorConfig` and pass it to the RE execute call

Reviewed By: stepancheg

Differential Revision: D50928319

fbshipit-source-id: 6347fc7fe66dd2d058b491853157ca8c8a3cb676