Move shed/more_futures to app/buck2_futures
Summary:
This generic shed library had grabbed dependencies on buck2_data, buck2_events etc. It stopped being generic a long time ago and became a core part of Buck2. Rename it appropriately.

Unfortunately DICE depends on buck2_futures, which is a bit of a mistake if we ever hope to open source DICE separately. This makes that clearer.

Reviewed By: ezgicicek

Differential Revision: D51538273

fbshipit-source-id: f6717f286a884127a415a72922ad47c2b3c9f178