add aast.bxl/bzl
Summary:
Now that koAST has been updated to be ready to run with `buck2`/`bxl` on remote execution, we can implement the `bxl` code to run it.

The way that this works is that the attributes of the regular target are duplicated into an anonymous `aast` target, with additional arguments added to run the plugin. There is currently some code here to support both jAST and "scan" mode, neither of which are properly implemented. This isn't a user-facing tool, it's an implementation detail of the Python CLI added in the next diff, and I don't anticipate that either of those will take long to implement, but I want to get koAST codemods working first.

The only really hacky part here, and the only part where I needed to touch the actual `buck2` Kotlin implementation, is that we can't have a `name` attribute on an anonymous target. So we need to support `_name` as well 🤷‍♂️. It's not //that// bad, but I definitely don't love it.

Reviewed By: kathryngray

Differential Revision: D51719025

fbshipit-source-id: 3f04eb06cf877f1afb245f35b07f3226d64d43fa