Add promise artifact mappings to AnalysisResult
Summary:
This stack is for fixing a race condition with promise artifacts.

# Problem

Currently, the caller of an anon_target creates a promise and fulfills it and the end of its analysis, but it can pass that promise to another anon_target and there can be a race where this second anon_target tries to access the artifact before the original (caller's) analysis is actually complete.

1. target T1 depends on anon target AT1
2. target T1 gets AT1's artifact A
3. target T1 passes that artifact to AT2

4. target T2 does the same steps 1-3.

At this point, whichever of T1 and T2 finish their initial analysis first will start triggering the anon target requests for the artifacts. Whichever of these targets gets to the AT2 request first will cause AT2 to refer to their instance of the A artifact. If something then causes the target to take a long time to totally finish their analysis (like that target depends on another anon target that takes a long time), then the other target (of T1 and T2) could finish and have something actually trigger building of the actions in AT2 that consume the A artifact, resulting in the promise not being fulfilled yet.

# Fix

We basically need to get rid of promises here. The ideal would be to move the rust `PromiseArtifact` into `ArtifactKind`, but that is a decently large refactor. The short term plan is to explicitly construct promised artifacts after evaluating the anon target, and pass these artifacts into the anon target key's analysis result. We will then use DICE to look up the results wherever we would try to access the `PromiseArtifact`'s artifact (getting the `OnceLock` value).

More details:

* In the anon target analysis, we will explicitly construct the artifacts by applying the anon target's own promise artifact lambda mappings to the anon target result (a provider collection)
* We will pass these explicitly constructed artifacts into the AnonTargetKey's AnalysisResult so that we can look it up via DICE later. We will pass it in as a `HashMap<PromiseArtifactId, Artifact>`. (This is what the current diff is doing)
* Anywhere that we call `run_promises`, we will assert the short paths and resolve the promised artifacts of the caller's analysis by calling DICE to look up the anon target key. During short path assertions, we currently also try to resolve the `PromiseArtifact` by checking that the artifact has been bound (`get_bound_artifact()`). Instead, we will need to use DICE to look up the caller's anon target's explicitly constructed artifacts, which is a blocking call, and should ensure that any dependent anon target analyses have finished
* The anon target caller will no longer perform the promise mapping (call to `StarlarkPromise::map`) on the anon target's unresolved promise, and the `PromiseArtifactRegistry` effectively just holds a list of declared promise artifacts of the caller
* We have to change `ArtifactGroup::assert_resolved()`, which is used to assert that an artifact has been resolved, to use DICE to look up the `AnonTargetKey`'s analysis result. This is an async call, so we have to update a few callsites to support that, notably some change to `BuildSignalsContext`

Reviewed By: cjhopman

Differential Revision: D52163174

fbshipit-source-id: cdde85b1854611c10c0c5d0b8532960bf1a12ecd