Add error diagnostics to ActionExecutionEnd
Summary: Used in the next diff to write the categories to `buck2_action_errors` so that we can get all the infra set up to support this intiative. I want to get these into `buck2_errors` as well as a set of categories across all actions ran for a single invocation, but that can come later.

Reviewed By: shonaganuma

Differential Revision: D51890532

fbshipit-source-id: be1a516bf826ff5f7c7388b353f39830b29683cf