Make ResolvedArtifactGroup::Artifact owned
Summary:
The rest of the stack starting from this diff contains all the changes needed to support using DICE to look up the resolved artifact for `ArtifactGroup::PromiseArtifact`. By the end of the stack, we will have introduced a new async `resolved_artifact()` function that calls DICE to look up the analysis result of the `AnonTargetKey` associated with the `ArtifactGroup::PromiseArtifact`, and all callsites of `assert_resolved()` are changed to use this function and also support async calls. `assert_resolved()` will be deprecated towards the end of the stack.

This diff makes `ResolvedArtifactGroup::Artifact` take an owned artifact instead of a reference, since anything looked up by DICE is owned, and you can't pass references of things you created within a function.

Reviewed By: cjhopman

Differential Revision: D52163168

fbshipit-source-id: 2fc01e85f80175d3504a6d49af986e0dd3324560