Revert D52163177: Multisect successfully blamed "D52163171: [buck2] Remove assert_resolved() in ensure_artifact_group_staged" for otest failure
Summary:
This diff is reverting D52163177
Depends on D52331769
D52163171: [buck2] Remove assert_resolved() in ensure_artifact_group_staged by wendy728 has been identified to be causing the following test failure:

Tests affected:
- [buck2/tests/e2e/build:test_uncategorized - test_uncategorized.py::test_keep_going](https://www.internalfb.com/intern/test/281475087544627/)

Here's the Multisect link:
https://www.internalfb.com/multisect/3828780
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: wendy728

Differential Revision: D52331775

fbshipit-source-id: 7571f9c51fba1c08068262ced93f0825dd539719