rules: Avoid `params.dep_link_strategy` more
Summary: The original description of D52450304 was actually a bit understated. As noted in a comment in that diff, that field is populated with a wrong-ish value for ReasonsTM, so we should avoid using that value if at all possible. This diff moves the `_handle_rust_artifact` invocation later, so that we can use the proper link strategy and not the sometimes-wrong one that's in `params`. This shouldn't be a behavioral change in practice because of `FORCE_RLIB` but it is closer to correct

Reviewed By: dtolnay

Differential Revision: D52499882

fbshipit-source-id: 9fd35d338f4886c46b4422dc9cbfecf3e77f9425