use buck config for the build mode to decide remote excution action key provider
Summary: The remote test execution action key provider is currently using fbcode//buck2/platform/build_mode:build_mode if the project is in a meta repo. However, not all meta repo has this target (e.g. whatsapp-android). It is better that we use a buck config to make it configurable.

Differential Revision: D52531004

fbshipit-source-id: 3ef58bf78ff9b077f31e7a739f258f0397426d84