Back out "buck2: add a buck2-only linker wrapper rule that preserves exe / flags difference"
Summary:
It looks like, with D52099825, this is no longer needed to unblock
building Go's stdlib from source.

It also appears that this leads to a bug with the `$(ldflags <target>)`
macro used in `cxx_genrule`, as it may mix flags constructed using the
exec-platform for `<target>` with in the `cxx_genrule` (which may not
match)
(https://www.internalfb.com/diff/D48899332?dst_version_fbid=1962894484105110&transaction_fbid=693256949613816).

Reviewed By: podtserkovskiy

Differential Revision: D52103735

fbshipit-source-id: 4d8faa715719b241b67e160c7b5353d7c6eefef5