Revert D52539366: Multisect successfully blamed "D52539366: [buck2] Adjust simpleconsole stderr keep alive timeout from 7s to 60s" for otest failure
Summary:
This diff is reverting D52539366
D52539366: [buck2] Adjust simpleconsole stderr keep alive timeout from 7s to 60s by milend has been identified to be causing the following test failure:

Tests affected:
- [buck2/tests/e2e/console:test_console - test_console.py::test_stale_snapshot](https://www.internalfb.com/intern/test/562950045941594/)

Here's the Multisect link:
https://www.internalfb.com/multisect/3902494
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: milend

Differential Revision: D52608314

fbshipit-source-id: 9bff079d916b2616c743c1f6e01276adbf5ed1a9