Add a `visit_field_with` method for custom control of field visitation.
Summary:
Adds the `visit_field_with` method which allows the user to control how a given visit for a field is visited.

The spirit of this methods is that it this lets the user use a custom, ad-hoc `Allocative` implementation for a given field, which is very helpful if the given type does not implement `Allocative` (from a third party library for instance. The function does nothing much special and the same pattern can be achieved manually using `enter` and `exit`, but with this it is more convenient and easier to get right.

To test that it works (and basically supersedes some of the other method's boilerplate) I reimplemented them with this function. This may be undesirable to you and I'm happy to revert those implementations if you wish.

X-link: https://github.com/facebookexperimental/allocative/pull/10

Reviewed By: ndmitchell

Differential Revision: D52630422

Pulled By: stepancheg

fbshipit-source-id: 3bcd10a31492c2a3d136a2adf4c83d32ef7de3ed