rules: Reimplement native unbundled deps to better support `FORCE_RLIB`
Summary:
The current implementation of native unbundled deps works by never using the Rust -> Rust link providers, instead using the native link providers even on Rust -> Rust edges. The result of this is that when a Rust binary depends on a library, building the binary ends up building the library both as a dylib and as an rlib - as an rlib because the Rust dependency handling requires that, and as a dylib because producing the link args for the Rust binary uses the regular, native version of the library's providers.

This diff removes that behavior, instead choosing to fix up the Rust -> Rust link providers in such a way that they include a correct "force rlib"-like representation of the Rust library

This is necessary because: 1. we need FORCE_RLIB to work, there's a reason it exist, 2. the alternative effectively disables pipelined builds, and so would be a major build speed loss, 3. the alternative results in static initializer duplication

Reviewed By: dtolnay

Differential Revision: D52499881

fbshipit-source-id: c3a889ac7abd67fa10b33c4ea140338111526958