buck2: http: add option to disable http2
Summary:
We have this SEV S382723

There are flaky tests because some http requests are failing

- Everstore data shows we are hitting no quota
- CDN data shows throttling is very low

But we see thousands of http requests failing, many at once. Which suggests that they were multiplexed on a single h2 connection that broke

Disabling h2 should tell us if that hunch is right. This diff adds that option

 ---

Most of this was non-creatively copied from the vpnless code.

NOTE: Not entirely sure the buckconfig setting is the way to go

Reviewed By: krallin

Differential Revision: D52683073

fbshipit-source-id: 046a2fd6b92a557fd3493525bb165c02bc2e2cb1