query: rewrite async_depth_first_postorder_traversal
Summary:
# Problem with query

We have an issue with query: it is single-threaded, and it is actually a problem.

But this is not easy to fix due to how our traits are designed:

- `AsyncTraversalDelegate` is both node visitor and query children
- It takes `&mut`
- Some implementation actually want `&mut` for the same field, like `somepath` implementation (this is dealt with in the following diff D52764092)

https://www.internalfb.com/code/fbsource/[4e2b834918e2796f2b133e5d3197a649fa9f70ba]/fbcode/buck2/app/buck2_query/src/query/traversal.rs?lines=39-49

Bottom line it is hard to parallelize query implementation.

I'm trying to clean up this code.

# struct Graph

This diff implements `Graph` data structure, which is basically the result of traversal with all nodes are stored in array, and all edges resolved to indices.

`async_fast_depth_first_postorder_traversal` was two staged function:
- unordered traversal to build node list (`async_unordered_traversal`)
- ordered traveral (`async_fast_depth_first_postorder_traversal`)

This diff rewrites both:
- first stage now builds `Graph` instance
- second stage now traverses using different function

# Results

Resulting code is somewhat faster (up to 15% faster query, see benchmark below). It is faster because second stage graph iteration is much faster now (code works synchrnously with indices instead of async callbacks and hashmaps).

The downside is both parts of the algorithm are copy-paste. I hope to get rid of such copy-paste later.

Reviewed By: wendy728

Differential Revision: D52761129

fbshipit-source-id: 98380efbf3aca4123c9174b087f4ec89fb695ea0