Add label parser to the Bazel LSP
Summary:
Instead of parsing labels ad-hoc in `resolve_folder`, this introduces a separate label parser. This allows re-using this when doing a "go to definition" for a label that is not in a load statement. See https://github.com/facebookexperimental/starlark-rust/issues/100 for more context.

X-link: https://github.com/facebookexperimental/starlark-rust/pull/101

Reviewed By: stepancheg

Differential Revision: D52887102

Pulled By: ndmitchell

fbshipit-source-id: e08c07b3d88c24d14030568dcc3ecf33ae8c243c