buck2: test: add the ability to specify a timeout
Summary:
In CI we need to enforce timeouts when running tests, but right now the
way that is done is by simply signalling the Buck2 command to exit by sending
it a SIGTERM (or maybe SIGINT but that does not matter).

The problem with doing that is that we can't trust that everything exited
cleanly. In particular, we don't know if the test runner will have had a chance
to report what we did vs. what we cancelled, report that we are done, etc.

As a result, right now those timeouts get handled by just retrying the whole
job (which often then times out again). Instead we'd like to experiment with
simply giving the command a time budget, and once that time budget is exceeded
we just stop.

This is what this does, you pass `--overall-timeout` to `buck2 test` and then
if we exceed the deadline that represents, we just pretend that you pressed
CTRL+C *but* wait for everything to actually finish cleanly (so by the time the
command exits, you should be able to expect that e.g. the test runner has
reported what you ran and that you can use this to e.g. trigger BRR).

Reviewed By: JakobDegen

Differential Revision: D52832187

fbshipit-source-id: 3162468544e23e2662cce40b7fe7cceeb3ea2021