AnonTargetAttrTraversal
Summary:
This is a bit unfortuante - we already have a `traverse()` function, but it takes in a `dyn ConfiguredAttrTraversal` which lives in the buck2_node crate, which doesn't have access to promise artifacts.

Adding a separate function seems like the easiest way to implement this - we just need to traverse the anon target attrs to get all the promise artifact attrs. Let me know if there are other suggestions, like moving the promise artifact attr into the buck2_node crate, or something like that

Reviewed By: cjhopman

Differential Revision: D52741553

fbshipit-source-id: bed794ba0b56c8ef0ad073fb0c3a9506a0055cb6