PromiseArtifactAttr
Summary:
There is another promise artifact race with cmd_args:

```
[2024-01-09T06:19:17.299-08:00] Action failed: anon//:_anon_link_rule@46a4dd0cf9f7b8f8 (write __link_groups__/libnodeapi.so.linker.argsfile)
[2024-01-09T06:19:17.299-08:00] tried to expand unresolved promise artifact `<promise artifact declared at fbcode/buck2/prelude/cxx/link.bzl:389:44-79 with short_path `libfolly_futures_tree.so.stub`>` to command line args, this should only happen after starlark objects are frozen and promises are resolved
```

The sharing that happens is when you have multiple targets (X, Y) that use anon target A's artifact, and passes them to anon target B. What happens is that B ends up getting the starlark promise artifact from one of X or Y, then either X or Y could finish first, and then anon target B's computations can start happening with these artifacts. If the flow is that B gets the artifact from X, but then X is really slow for some reason, and Y finishes but X hasn't finished, then B's artifacts can start being requested. The problem is the OnceLock artifact came from X and wasn't fulfilled yet.

We can solve this by preventing this sharing from the first place. The anon target key encodes the promise artifact in a way that only holds data that doesn't depend on X or Y fulfilling it. Currently, the anon target attr accepts the entire StarlarkPromiseArtifact. In coercion, it's still a starlark value, and then goes through resolution (and we don't do anything fancy to the starlark value).

Plan: change `AnonTargetAttr::PromiseArtifact` to use `PromiseArtifactAttr` instead of the real `StarlarkPromiseArtifact` from the consumer analysis. The end state is that we construct a new fulfilled promise artifact during this anon target's analysis so that things aren't shared in a way that can cause a race condition.

The last diff in the stack pretty much shows all the pieces coming together to achieve this

Reviewed By: cjhopman

Differential Revision: D52741549

fbshipit-source-id: 4f25512c2b7461ad0136d4bc0274f5cc1bbba814