Revert D52877026: Multisect successfully blamed "D52877026: delay site import" for otest failure
Summary:
This diff is reverting D52877026
D52877026: delay site import by zsol has been identified to be causing the following test failure:

Tests affected:
- [tools/make_par:environment_tests - test_ld_preload_xar_forkserver (tools.make_par.environment_tests.EnvironmentTests)](https://www.internalfb.com/intern/test/281475066289502/)

Here's the Multisect link:
https://www.internalfb.com/multisect/4026549
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: zsol

Differential Revision: D52906644

fbshipit-source-id: d26c06e22921a4337f35572a73fcabf7dfad84bb