Advertise query docs in help
Summary:
TIL we publish generate query documentation.

Put the link in `buck2 xquery --help` to make it better discoverable.

Code is pulled out of `///` doc because clap does not understand `#[doc]` comments, and conditionally enable-disable rustdoc comments in internal-oss is hard.

Comments are also reformatted because
- without verbatim clap does not produce comments I like
- with verbatim reformatting is needed

Text width is 72, that is perfect text width, and that is what `fmt` does, which I used to reformat the comments.

Reviewed By: wendy728

Differential Revision: D52905040

fbshipit-source-id: 0f9419bbfde335a3884b93f14717646f2eb0313d