quick fix for shlex deprecations
Summary:
`shlex::quote` and `shlex::join` are deprecated in `shlex` 1.3.0 version. In fbcode we use 1.0.0 but in OSS we use 1.3.0. See failed job https://www.internalfb.com/sandcastle/workflow/9007199268894132?entry_point=5
This is a quick fix to unblock CI.
The proper fix is to update `shlex` in fbcode to 1.3.0 and fix all errors, which I'm going to do next.

Reviewed By: stepancheg

Differential Revision: D52959838

fbshipit-source-id: c4ad7ac8787663ddb3a53ea22ba9308e363800e7