Rewrite somepath to avoid building whole dep tree
Summary:
Use async BFS to traverse the graph without building the whole universe.

There are two reasons for this:
- if dependency tree is large, but the resulting path is short, it will be much faster
- if unconfigured dependency graph is broken somewhere down (like today on mac), we should not fail

The drawback is, in certain situations, where large portion of universe is needed, `somepath` will be slower.

Another drawback is the implementation is quite complicated. It can be made easier if we were OK with extra work, which I tried to avoid. For example, we avoid queueing child node if it is already a match by node key.

Reviewed By: wendy728

Differential Revision: D52897493

fbshipit-source-id: 77b3e6defab82b7d45e7d7e359956702b1744016