Only xz-compress native libs in modules if base apk does the same
Summary:
As in the title. We will get rid of this compression option (with XZ) altogether in the last diff on the stack.

The aim is to remove the concatenation hack + XZ compression, and rely on superpack only to compress everything for us. Otherwise, asset libraries will be shipped uncompressed.

Since D52296279, all major apps should have voltron libs compression turned on.

Without this change, once we switch from xzso to noxz/xzdex, native libraries in voltron modules will still be packaged as libs.xzs. We don't want that, since with D52389378 we're expecting individual libraries to run superpack on top of.

Reviewed By: IanChilds

Differential Revision: D53000687

fbshipit-source-id: 4844e24cfeecf4be46166f24f513420ae678d565