rules: Hide the rlibs from rustc
Summary: The underlying observation here is that `native_unbundle_deps` has already done all the hard work associated with getting link groups working; Rust dependencies are available in the usual form on the build graph at link time. So if we were to just turn on link groups now, it would almost work; the only problem is that rustc would keep passing the rlibs to the linker, which appear before the link groups in the link line, and so get linked in and negate all the benefits of link groups. Using hollow rlibs prevents this from being a problem

Reviewed By: dtolnay

Differential Revision: D52537370

fbshipit-source-id: 7ca11f7a4b0d208ebe1c4eb47b78f1e6e195c389