test_new_error_handler for testing error handler impl
Summary:
I want to provide a way for rule authors to test their action error handler impls easily without needing to actually run a failed action.

This stack will expose some functions for action error handler starlark types, and the eventual plan is for rule authors to use `bxl_test` to easily test their stuff. An example test is in the last diff of the stack (I do want to provide a synchronous ensure function to get rid of the need for dynamic outputs, but that can come later).

Reviewed By: stepancheg

Differential Revision: D53098694

fbshipit-source-id: 5ea83ca8b4a889e08bd3e9ad1207aadb1bdfc924