Implement [swift-interface] subtarget on apple_library's
Summary: We want to be able to generate `.swift` interface files for general `apple_library`s (particularly ones that don't include any `.swift` files). In order to achieve that we reuse the `swiftc` flag calculation, as if this library was a 'mixed' `apple_library` (i.e. it contain both `.swift` and non-`.swift` source files, so the current target is considered an 'underlying module') and pass those arguments (after translation) to `swift-ide-test`.

Reviewed By: maxovtsin

Differential Revision: D53098987

fbshipit-source-id: 155b52db0dcfa693a0ee07c6d61b81c899496385