Implement [swift-interface] subtarget on apple_library's
Summary:
We want to be able to generate `.swift` interface files for general `apple_library`s (particularly ones that don't include any `.swift` files). In order to achieve that we reuse the `swiftc` flag calculation, as if this library was a 'mixed' `apple_library` (i.e. it contain both `.swift` and non-`.swift` source files, so the current target is considered an 'underlying module') and pass those arguments (after translation) to `swift-ide-test`.

The previous attempt at this ended up in double declaring the swift modulemap artifact. This approach avoids that shortcoming by reusing the flag calculation for both swift compilation artifacts & the `[swift-interface]` artifact.

Reviewed By: maxovtsin

Differential Revision: D53203807

fbshipit-source-id: d1dd1eb13abc72126a3affaecd4850b1fb7a61ed