buck2: test: simplify exit code handling
Summary:
The current implementation is unnecessarily complicated: if there are build
failures, the server withholds the exit code from Tpx, and then the client
infers that if there is an exit code from Tpx we should use it, otherwise there
must be a build failure, and then we should produce an exit code for that.

The awkward bit here is that this is actually slightly incorrect, since if Tpx
did not provide an exit code, we would exit with zero.

This diff fixes that so that instead, we give the client all the info, and let
it decide what exit code to produce. This also fixes the fact that right now,
if the overall timeout is hit, then we do withhold the exit code from Tpx, and
as a result we show an exit code 3.

Note that it's not actually possible for this field to be null. The reason
it's not possible is because this field is unconditionally set in end_of_tests:

https://www.internalfb.com/code/fbsource/[3dd71c7a711a131286fc84d6d0aaf670b1adc371]/fbcode/buck2/app/buck2_test/src/orchestrator.rs?lines=422-427

And if that method is not called (and the channel is not closed), then we
inject an error into that channel:

https://www.internalfb.com/code/fbsource/[57b2f8d2755901e433016fabca59e62ba7e0d727]/fbcode/buck2/app/buck2_test/src/command.rs?lines=582-584

Which in turn causes this to error out:

https://www.internalfb.com/code/fbsource/[57b2f8d2755901e433016fabca59e62ba7e0d727]/fbcode/buck2/app/buck2_test/src/command.rs?lines=540-546

Reviewed By: lmvasquezg

Differential Revision: D53266108

fbshipit-source-id: 612ce67820e6bb31362b406dff1104644cc5d527