manifold_client_wrapper: add preferred_linkage="static"
Summary:
~~excluding dependencies of omnibus roots because they don't have preferred linkage "static" results in rust cxx artifacts not making it into 'libomnibus.so' in the next diff up the stack.~~

~~it's not clear why that restriction has existed before now. maybe andrewjcg or ndmitchell know or maybe it was just copied over from buck1 (and the rationale not known)?~~

`manifold_client_wrapper` is a C++ dependency of `manifold_client` which is ultimately a dependency of `container_manifest_lib_py`.

since `manifold_client_wrapper` is not marked static, it ends up in the 'libomnibus.so' produced for the test `fbcode//fblearner/flow/projects/modelstore/tests:pyper_publish_integration_test` of which `container_manifest_lib_py` is an omnibus root.

`manifold_client` and `manifold_client-bridge` are preferred linkage static and are therefore linked directly into `container_manifest_lib_py` and  excluded from this 'libomnibus.so'.

the result of all this is link errors in 'libomnibus.so'.

marking `manifold_client_wrapper` as preferred linkage static ensures now that it is linked directly into `container_manifest_py` (as `manifold_client` and `manifold_client-bridge` are) and no longer considered for linking into 'libomnibus.so'.

Reviewed By: dtolnay

Differential Revision: D53275848

fbshipit-source-id: f68f916b30e07ebbbfc6eb8ed80d22a6ae6cd825