Group all functions of buck2_error::Error
Summary:
This is for code navigation. When code is scattered among three files, it is hard to see what are the functions.

At least in CLion. Suppose a user needs to find `tags` function (real story). A user needs to click through three files to find them:

{F1427645242}

Now a user can just click outline, and everything's here.

{F1427650144}

I don't know how bad it is with other IDEs.

Reviewed By: JakobDegen

Differential Revision: D53326411

fbshipit-source-id: b0f687014e3995978b28396e49a93f0988806920