configurations: Cache exec platform resolution on dice
Summary:
There's a big comment here that explains why putting exec platform resolution behind a dice key is a good idea:

https://www.internalfb.com/code/fbsource/[e8506406688a]/fbcode/buck2/app/buck2_configured/src/configuration/calculation.rs?lines=375

But then the code doesn't actually do that. So let's.

This is perf neutral for max RSS and time again, but for steady state memory it's a small improvment:

- Allocated: 5455745432 -> 5410306552 (0.9%)
- Resident:  7263719424 -> 7177924608 (1.2%)

This isn't a huge win, but one of the side benefits of this is that - since I'm planning to add some logic to this code soon - this gives me piece of mind that I can do that in a perf neutral way

Reviewed By: scottcao

Differential Revision: D52654693

fbshipit-source-id: c5d8dbb26fd9330d973bdbbc8cfcdaaab135a44d