Classify server stderr on gRPC error
Summary:
If it is gRPC error, assume it is bas enough, and start classifying server stderr, even if server maybe not dead.

Technically we could actually check if server is alive, but since this does not happen often (at least, it should not happen often), it is safer to assume server is dead. And classify stderr.

For now have four tags extracted from stderr, and attached to the gRPC error:
- stderr is empty
- unknown
- panic
- stack overflow

The last one is the main reason for this stack: I'd like to know, how often stack overflow happens in buck2. User [reported stack overflow](https://fb.workplace.com/groups/buck2users/posts/1363623130961610), but we have no place to count it.

Reviewed By: JakobDegen

Differential Revision: D53338307

fbshipit-source-id: 0d0a3a58ab0a99c900c14d4d2d242c95df3c1ba8