Best error tag
Summary:
Panic is more interesting than watchman error, and watchman error is more interesting than starlark error.

Pick the best tag and store it in the invocation record, with intention to propagate it to Scuba (D53342062).

Ideally, this classification should live in ingress. But C++ violates Universal Declaration of Human Rights, article 5:

> No one shall be subjected to torture or to cruel, inhuman or degrading treatment or punishment.

so until we rewrite ingress in Rust, I would prefer to avoid it.

Bikeshedding for field names are welcome. "Best" is probably not the best, but it is the shortest. Otherwise I'd call it "most interesting".

Reviewed By: JakobDegen

Differential Revision: D53341163

fbshipit-source-id: e97f3927a735bf213fd9dcbcc5081b68218ba9c3