Add _kwargs
Summary: I want to pass `rule_name` parameter to `cfg_constructor_pre_constraint_analysis` function, but now that modifiers are rolled out, we need to make sure we don't break starlark when we make core changes to cfg constructor impl. Adding `**kwargs` in this diff as a parameter to `cfg_constructor_pre_constraint_analysis`  to ensure this is the case.

Reviewed By: JakobDegen

Differential Revision: D53380256

fbshipit-source-id: cb42a01d8250902138e5d6c971b9cb60cb583ee1