buck2_execute: recalculate relative symlink targets for RE on Windows
Summary:
`buck-out` can be a symlink on local host but on RE it's always a regular folder.

Because of that relative targets should be different on Windows.

On Windows relative symlinks are relative to canonical path of symlink (real path after all symlinks in path are resolved) and it doesn't matter via which path they are accessed.

To make such symlinks work on RE, we need to recalculate target of the symlink.

In opposite way (when symlinks are created on RE and materialized locally) we don't have such issue because all symlinks are materialized as actual files which symlink was pointing to.

Reviewed By: wendy728

Differential Revision: D53429335

fbshipit-source-id: ee14a46ba6955301fd26a43208f45d64fd5c17f8