eval_single_query from eval_multi_query (fix universe bug)
Summary:
Just invoke single queries independently, instead of evaluating all the queries in singe environment.

This is necessary for `cquery`, but for `uquery` this is unnecessary work: common literals needs to be resolved multiple time. I think this is fine, but FYI.

This fixes cquery universe bug (described in D53110819).

Reviewed By: iguridi

Differential Revision: D53112680

fbshipit-source-id: f02d958f2e38df5f6becc5d8779bfc923dbf6d42